+++ tests/Fixer/Strict/DeclareStrictTypesFixerTest.php
@@ -24,10 +24,6 @@ use PhpCsFixer\WhitespacesFixerConfig;
  */
 final class DeclareStrictTypesFixerTest extends AbstractFixerTestCase
 {
-    private static $configurationAddMissing = ['add_missing' => true];
-    private static $configurationRelocateNextLine = ['relocate_to' => 'next'];
-    private static $configurationRelocateSameLine = ['relocate_to' => 'same'];
-
     /**
      * @param string      $expected
      * @param null|string $input
@@ -118,241 +114,6 @@ $a = 456;
 declare(strict_types=1);
 /**/',
                 '<?php /**/',
-            ],
-        ];
-    }
-
-    /**
-     * @param string      $expected
-     * @param null|string $input
-     * @param null|array  $configuration
-     *
-     * @dataProvider provideNextLineFixCases
-     * @requires PHP 7.0
-     */
-    public function testNextLineFix($expected, $input = null, array $configuration = null)
-    {
-        if (null !== $configuration) {
-            $this->fixer->configure($configuration);
-        }
-
-        $this->doTest($expected, $input);
-    }
-
-    public function provideNextLineFixCases()
-    {
-        return [
-            [
-                '<?php
-declare/* A b C*/(strict_types=1);
-declare(ticks=1);
-//
-
-
-namespace A\B\C;
-class A {
-}',
-                '<?php
-declare(ticks=1);
-//
-declare/* A b C*/(strict_types=1);
-
-namespace A\B\C;
-class A {
-}',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-declare/* A b C*/(strict_types=1);',
-                null,
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-deClarE(strict_types=1)
-/**/ /**/       ?>Test',
-                '<?php /**/ /**/ deClarE  (STRICT_TYPES=1)    ?>Test',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-DECLARE(strict_types=1);
-                       ',
-                '<?php            DECLARE  (    strict_types=1   )   ;',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-declare(strict_types=1);
-                /**/
-                ',
-                '<?php
-                /**/
-                declare(strict_types=1);',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-declare(strict_types=1);
-
-                phpinfo();',
-                '<?php
-
-                phpinfo();',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-declare(strict_types=1);
-
-/**
- * Foo
- */
-phpinfo();',
-                '<?php
-
-/**
- * Foo
- */
-phpinfo();',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-declare(strict_types=1);
-phpinfo();',
-                '<?php phpinfo();',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-declare(strict_types=1);
-$a = 456;
-',
-                '<?php
-$a = 456;
-',
-                self::$configurationRelocateNextLine,
-            ],
-            [
-                '<?php
-declare(strict_types=1);
-/**/',
-                '<?php /**/',
-                self::$configurationRelocateNextLine,
-            ],
-        ];
-    }
-
-    /**
-     * @param string      $expected
-     * @param null|string $input
-     * @param null|array  $configuration
-     *
-     * @dataProvider provideSameLineFixCases
-     * @requires PHP 7.0
-     */
-    public function testSameLineFix($expected, $input = null, array $configuration = null)
-    {
-        if (null !== $configuration) {
-            $this->fixer->configure($configuration);
-        }
-
-        $this->doTest($expected, $input);
-    }
-
-    public function provideSameLineFixCases()
-    {
-        return [
-            [
-                '<?php declare/* A b C*/(strict_types=1);
-declare(ticks=1);
-//
-
-
-namespace A\B\C;
-class A {
-}',
-                '<?php declare(ticks=1);
-//
-declare/* A b C*/(strict_types=1);
-
-namespace A\B\C;
-class A {
-}',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php declare/* A b C*/(strict_types=1);',
-                null,
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php deClarE(strict_types=1)
-/**/ /**/       ?>Test',
-                '<?php /**/ /**/ deClarE  (STRICT_TYPES=1)    ?>Test',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php DECLARE(strict_types=1);
-                       ',
-                '<?php            DECLARE  (    strict_types=1   )   ;',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php declare(strict_types=1);
-                /**/
-                ',
-                '<?php
-                /**/
-                declare(strict_types=1);',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php declare(strict_types=1);
-
-                phpinfo();',
-                '<?php
-
-                phpinfo();',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php declare(strict_types=1);
-
-/**
- * Foo
- */
-phpinfo();',
-                '<?php
-
-/**
- * Foo
- */
-phpinfo();',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php declare(strict_types=1);
-phpinfo();',
-                '<?php phpinfo();',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php declare(strict_types=1);
-$a = 456;
-',
-                '<?php
-$a = 456;
-',
-                self::$configurationRelocateSameLine,
-            ],
-            [
-                '<?php declare(strict_types=1);
-/**/',
-                '<?php /**/',
-                self::$configurationRelocateSameLine,
             ],
         ];
     }
